requirements.txt
streamlit
pandas
plotly
python-telegram-bot
requests
numpy
matplotlib
ta
